requirements.txt
aioredis==1.3.1
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.3
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
celery==5.1.2
certifi==2020.6.20
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect==1.1.0
daphne==3.0.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-braces==1.14.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.3.0
django-filter==2.3.0
django-heroku==0.3.1
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-timezone-field==4.2.1
djangorestframework==3.11.0
djangorestframework-queryfields==1.0.0
drf-yasg==1.20.0
fake-useragent==0.1.11
fcm-django==0.3.10
flake8==3.9.2
gunicorn==20.0.4
hiredis==2.0.0
hyperlink==21.0.0
idna==2.9
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonfield==3.1.0
kombu==5.1.0
lxml==4.6.3
MarkupSafe==2.0.1
mccabe==0.6.1
MechanicalSoup==1.1.0
msgpack==1.0.2
oauthlib==3.1.0
packaging==21.0
parse==1.19.0
Pillow==8.3.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyee==8.1.0
pyfcm==1.4.9
pyflakes==2.3.1
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-html==0.10.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
service-identity==21.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.2.1
sqlparse==0.3.1
toml==0.10.1
tqdm==4.62.0
tweepy==3.10.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.25.9
validate-email==1.3
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==8.1
whitenoise==5.1.0
zope.interface==5.4.0
